store: Move subscription handling from Store to SubscriptionManager
Store continues to be a facade for subscribing
