+++ store/postgres/src/store_events.rs
@@ -1,3 +1,8 @@
+use futures::sync::mpsc::{channel, Sender};
+use std::collections::HashMap;
+use std::sync::{Arc, Mutex, RwLock};
+use uuid::Uuid;
+
 use crate::notification_listener::{NotificationListener, SafeChannelName};
 use graph::prelude::serde_json;
 use graph::prelude::*;
@@ -42,5 +47,133 @@ impl EventProducer<StoreEvent> for StoreEventListener {
                 }))
             },
         )
+    }
+}
+
+pub(crate) struct SubscriptionManager {
+    logger: Logger,
+    subscriptions: Arc<RwLock<HashMap<String, Sender<StoreEvent>>>>,
+
+    /// listen to StoreEvents generated when applying entity operations
+    listener: Mutex<StoreEventListener>,
+}
+
+impl SubscriptionManager {
+    pub fn new(logger: Logger, postgres_url: String) -> Self {
+        let mut listener = StoreEventListener::new(&logger, postgres_url);
+        let store_events = listener
+            .take_event_stream()
+            .expect("Failed to listen to entity change events in Postgres");
+
+        let manager = SubscriptionManager {
+            logger,
+            subscriptions: Arc::new(RwLock::new(HashMap::new())),
+            listener: Mutex::new(listener),
+        };
+
+        // Deal with store subscriptions
+        manager.handle_store_events(store_events);
+        manager.periodically_clean_up_stale_subscriptions();
+
+        let mut listener = manager.listener.lock().unwrap();
+        listener.start();
+        drop(listener);
+
+        manager
+    }
+
+    /// Receive store events from Postgres and send them to all active
+    /// subscriptions. Detect stale subscriptions in the process and
+    /// close them.
+    fn handle_store_events(
+        &self,
+        store_events: Box<dyn Stream<Item = StoreEvent, Error = ()> + Send>,
+    ) {
+        let logger = self.logger.clone();
+        let subscriptions = self.subscriptions.clone();
+
+        graph::spawn(
+            store_events
+                .for_each(move |event| {
+                    let senders = subscriptions.read().unwrap().clone();
+                    let logger = logger.clone();
+                    let subscriptions = subscriptions.clone();
+
+                    // Write change to all matching subscription streams; remove subscriptions
+                    // whose receiving end has been dropped
+                    stream::iter_ok::<_, ()>(senders).for_each(move |(id, sender)| {
+                        let logger = logger.clone();
+                        let subscriptions = subscriptions.clone();
+
+                        sender.send(event.clone()).then(move |result| {
+                            match result {
+                                Err(_send_error) => {
+                                    // Receiver was dropped
+                                    debug!(logger, "Unsubscribe"; "id" => &id);
+                                    subscriptions.write().unwrap().remove(&id);
+                                    Ok(())
+                                }
+                                Ok(_sender) => Ok(()),
+                            }
+                        })
+                    })
+                })
+                .compat(),
+        );
+    }
+
+    fn periodically_clean_up_stale_subscriptions(&self) {
+        use futures03::stream::StreamExt;
+
+        let logger = self.logger.clone();
+        let subscriptions = self.subscriptions.clone();
+
+        // Clean up stale subscriptions every 5s
+        graph::spawn(
+            tokio::time::interval(Duration::from_secs(5)).for_each(move |_| {
+                let mut subscriptions = subscriptions.write().unwrap();
+
+                // Obtain IDs of subscriptions whose receiving end has gone
+                let stale_ids = subscriptions
+                    .iter_mut()
+                    .filter_map(|(id, sender)| match sender.poll_ready() {
+                        Err(_) => Some(id.clone()),
+                        _ => None,
+                    })
+                    .collect::<Vec<_>>();
+
+                // Remove all stale subscriptions
+                for id in stale_ids {
+                    debug!(logger, "Unsubscribe"; "id" => &id);
+                    subscriptions.remove(&id);
+                }
+
+                futures03::future::ready(())
+            }),
+        );
+    }
+
+    pub fn subscribe(&self, entities: Vec<SubgraphEntityPair>) -> StoreEventStreamBox {
+        let subscriptions = self.subscriptions.clone();
+
+        // Generate a new (unique) UUID; we're looping just to be sure we avoid collisions
+        let mut id = Uuid::new_v4().to_string();
+        while subscriptions.read().unwrap().contains_key(&id) {
+            id = Uuid::new_v4().to_string();
+        }
+
+        debug!(self.logger, "Subscribe";
+               "id" => &id,
+               "entities" => format!("{:?}", entities));
+
+        // Prepare the new subscription by creating a channel and a subscription object
+        let (sender, receiver) = channel(100);
+
+        // Add the new subscription
+        let mut subscriptions = subscriptions.write().unwrap();
+        subscriptions.insert(id, sender);
+
+        // Return the subscription ID and entity change stream
+        StoreEventStream::new(Box::new(receiver)).filter_by_entities(entities)
     }
 }
